requirements.txt
loguru~=0.7.2
pybit~=5.7.0
python-dotenv~=1.0.0
aiohttp~=3.9.4
requests~=2.31.0
aiogram~=3.12.0
web3==6.20.2
pycryptodome==3.18.0
uniswap-python==0.7.1

